--- src/components/SparcTooltip/index.js
@@ -1,8 +1,0 @@
-import SparcTooltip from './src/SparcTooltip';
-
-/* istanbul ignore next */
-SparcTooltip.install = function(Vue) {
-  Vue.component(SparcTooltip.name, SparcTooltip);
-};
-
-export default SparcTooltip;
