+++ src/components/Tooltip/index.js
@@ -1,0 +1,8 @@
+import Tooltip from './src/Tooltip';
+
+/* istanbul ignore next */
+Tooltip.install = function(Vue) {
+  Vue.component(Tooltip.name, Tooltip);
+};
+
+export default Tooltip;
