+++ src/components/index.js
@@ -13,7 +13,7 @@ import TwelveLaboursFooter from './TwelveLaboursFooter'
 import TabNav from './TabNav'
 import Pagination from './Pagination'
 import PaginationMenu from './PaginationMenu'
-import SparcTooltip from './SparcTooltip'
+import Tooltip from './Tooltip'
 
 import '@/assets/icons'
 
@@ -25,7 +25,7 @@ const Components = {
   TabNav,
   Pagination,
   PaginationMenu,
-  SparcTooltip
+  Tooltip
 }
 
 Object.keys(Components).forEach(component => {
