+++ src/stories/tooltip/tooltip.stories.mdx
@@ -13,9 +13,9 @@ The item for which the tooltip will be connected with should use the `item` slot
 
 #### Code
 ```html
-<sparc-tooltip placement="top-center" content="top-center">
+<tooltip placement="top-center" content="top-center">
   <el-button slot="item">Hover over me!</el-button>
-</sparc-tooltip>
+</tooltip>
 ```
 
 ## Content
@@ -26,9 +26,9 @@ The content to be displayed in the tooltip can be customized through the `conten
 #### Code
 
 ```html
-<sparc-tooltip placement="top-center" content="Sample content text">
+<tooltip placement="top-center" content="Sample content text">
   <el-button slot="item">Hover over me!</el-button>
-</sparc-tooltip>
+</tooltip>
 ```
 
 ### More Content
@@ -40,10 +40,10 @@ Multi lined content can be displayed in the tooltip by using the `data` slot.
 #### Code
 
 ```html
-<sparc-tooltip placement="top-center">
+<tooltip placement="top-center">
   <div slot="data">This is multi-line<br/>content sample text</div>
   <el-button slot="item">Hover over me!</el-button>
-</sparc-tooltip>
+</tooltip>
 ```
 
 ## Placement
@@ -54,19 +54,19 @@ Tooltips can be in different layouts based on the position of the arrow.
 #### Code
 
 ```html
-<sparc-tooltip :placement="top-left" content="top-left">
+<tooltip :placement="top-left" content="top-left">
   <el-button slot="item">top-left</el-button>
-</sparc-tooltip>
+</tooltip>
 
-<sparc-tooltip :placement="top-center" content="top-center">
+<tooltip :placement="top-center" content="top-center">
   <el-button slot="item">top-center</el-button>
-</sparc-tooltip>
+</tooltip>
 
 ...
 
-<sparc-tooltip :placement="right-bottom" content="right-bottom">
+<tooltip :placement="right-bottom" content="right-bottom">
   <el-button slot="item">right-bottom</el-button>
-</sparc-tooltip>
+</tooltip>
 ```
 
 #### Directions
@@ -94,8 +94,8 @@ Content with links will appear as follows.
 
 #### Code
 ```html
-<sparc-tooltip placement="top-center">
+<tooltip placement="top-center">
   <div slot="data">Content with <a href="#">link</a></div>
   <el-button slot="item">Hover over me!</el-button>
-</sparc-tooltip>
+</tooltip>
 ```
